requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
# IA Core & Azure SDKs
openai>=1.0.0
semantic-kernel==1.18.0
azure-identity==1.15.0
azure-search-documents==11.6.0b1
azure-cosmos==4.5.1
azure-ai-contentsafety==1.0.0
azure-cognitiveservices-speech==1.37.0
azure-ai-textanalytics==5.3.0
azure-mgmt-automation==1.0.0
azure-storage-blob==12.19.0
azure-ai-documentintelligence==1.0.2
# Datos y Utilidades
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
httpx==0.26.0
tqdm==4.66.0